+++ Changelog.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented here.
 
 ## [unreleased]
 - Add tidyverse as a default R tester package (#512)
+
+## [v2.4.3]
 - Omit skipped test cases in Python tester (#522)
 
 ## [v2.4.2]